LayoutTests/http/tests/security/contentSecurityPolicy/object-src-none-blocked-expected.txt

CONSOLE MESSAGE: Refused to load object from 'data:application/x-webkit-test-netscape,alertwhenloaded' because of Content-Security-Policy.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
